backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-engineio==4.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
